Fix Gtest Warnings

+1: simbot

Merged-on: https://assembla.com/code/OpenSurgSim/git/merge_requests/4134243

+++ SurgSim/Collision/UnitTests/SegmentMeshTriangleMeshContactCalculationTests.cpp
@@ -1,5 +1,5 @@
 // This file is a part of the OpenSurgSim project.
-// Copyright 2013, SimQuest Solutions Inc.
+// Copyright 2013-2016, SimQuest Solutions Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -167,7 +167,7 @@ protected:
 
 	void testSegmentMeshTriangleMeshDCD(std::string scenario, std::shared_ptr<SegmentMeshShape> segmentMeshShape,
 		RigidTransform3d segmentMeshShapeTransform, std::shared_ptr<MeshShape> meshShape,
-		RigidTransform3d meshShapeTransform, int expectedNumContacts = 0)
+		RigidTransform3d meshShapeTransform, size_t expectedNumContacts = 0)
 	{
 		for (const auto& transform : m_transforms)
 		{
@@ -222,7 +222,7 @@ protected:
 				calcContact.calculateContact(pair2);
 
 				// There should be no intersections.
-				EXPECT_EQ(0, pair2->getContacts().size());
+				EXPECT_EQ(0u, pair2->getContacts().size());
 			}
 		}
 	}
